+++ CHANGELOG.md
@@ -2,13 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-
+###
 ## [1.0.1] / 2026-xx-xx
+
 ### Added
+- Plugin window can be maximized/minimized
+- Settings window can be resized
+
 ### Fixed
 - Sentences with "!" as first character are accepted.  
 
-###
 
+###
 ## [1.0.0] / 2026-04-15
 ### Initial release