+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ###
-## [1.0.1] / 2026-xx-xx
+## [v1.0.1] / 2026-xx-xx
 
 ### Added
 - Plugin window can be maximized/minimized
@@ -14,5 +14,5 @@ All notable changes to this project will be documented in this file.
 
 
 ###
-## [1.0.0] / 2026-04-15
+## [v1.0.0] / 2026-04-15
 ### Initial release